refactor: reorganize headers and implementations
