+++ include/nuri/core/graph/traits.h
@@ -1,0 +1,16 @@
+//
+// Project NuriKit - Copyright 2025 SNU Compbio Lab.
+// SPDX-License-Identifier: Apache-2.0
+//
+
+#ifndef NURI_CORE_GRAPH_TRAITS_H_
+#define NURI_CORE_GRAPH_TRAITS_H_
+
+namespace nuri {
+namespace internal {
+  template <class GT>
+  struct GraphTraits;
+}
+}  // namespace nuri
+
+#endif /* NURI_CORE_GRAPH_TRAITS_H_ */
