+++ python/src/nuri/core/substructure.cpp
@@ -20,9 +20,9 @@
 #include <pybind11/pytypes.h>
 
 #include "nuri/eigen_config.h"
-#include "nuri/core/graph.h"
+#include "nuri/core/container/index_set.h"
+#include "nuri/core/container/property_map.h"
 #include "nuri/core/molecule.h"
-#include "nuri/core/property_map.h"
 #include "nuri/python/core/containers.h"
 #include "nuri/python/core/core_module.h"
 #include "nuri/python/utils.h"
