+++ src/algo/guess/3d.cpp
@@ -24,9 +24,10 @@
 #include "nuri/eigen_config.h"
 #include "nuri/algo/guess.h"
 #include "nuri/algo/rings.h"
+#include "nuri/core/container/index_set.h"
 #include "nuri/core/element.h"
 #include "nuri/core/geometry.h"
-#include "nuri/core/graph.h"
+#include "nuri/core/graph/graph.h"
 #include "nuri/core/molecule.h"
 #include "nuri/utils.h"
