+++ src/fmt/mol2.cpp
@@ -35,10 +35,9 @@
 #include "nuri/eigen_config.h"
 #include "fmt_internal.h"
 #include "nuri/algo/guess.h"
+#include "nuri/core/container/property_map.h"
 #include "nuri/core/element.h"
-#include "nuri/core/graph.h"
 #include "nuri/core/molecule.h"
-#include "nuri/core/property_map.h"
 #include "nuri/fmt/base.h"
 #include "nuri/utils.h"
