+++ src/fmt/pdb.cpp
@@ -36,10 +36,10 @@
 
 #include "nuri/eigen_config.h"
 #include "fmt_internal.h"
+#include "nuri/core/container/property_map.h"
 #include "nuri/core/element.h"
-#include "nuri/core/graph.h"
+#include "nuri/core/graph/graph.h"
 #include "nuri/core/molecule.h"
-#include "nuri/core/property_map.h"
 #include "nuri/fmt/base.h"
 #include "nuri/utils.h"
