+++ test/core/graph/graph_test.cpp
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "nuri/core/graph.h"
+#include "nuri/core/graph/graph.h"
 
 #include <iterator>
 #include <type_traits>
